Updated Book Entity, Added A book Controller to display book details, and cart page, added css stylesheets

+++ src/main/java/org/example/BookRepository.java
@@ -2,10 +2,10 @@ package org.example;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface BookRepository extends CrudRepository<Book, Long> {
+public interface BookRepository extends CrudRepository<Book, Integer> {
 
-    Book findById(long id);
-    void deleteById(long id);
+    Book findById(int id);
+    void deleteById(int id);
     Iterable<Book> findAll();
 
 }
